Restructured files to conform to task specification
